updated SITT count monitor
