+++ pylabnet/hardware/counter/swabian_instruments/cnt_monitor.py
@@ -93,89 +93,3 @@ class Wrap:
         # Set attribute to validated ch_list
         self._ch_list = ch_list
 
-
-class Service(ServiceBase):
-
-    def exposed_start_counting(self, bin_width=1000000000, n_bins=10000):
-        return self._module.start_ctr(
-            bin_width=bin_width,
-            n_bins=n_bins
-        )
-
-    def exposed_clear_counter(self):
-        return self._module.clear_ctr()
-
-    def exposed_get_counts(self):
-        res_pickle = self._module.get_counts()
-        return pickle.dumps(res_pickle)
-
-    def exposed_get_x_axis(self):
-        res_pickle = self._module.get_x_axis()
-        return pickle.dumps(res_pickle)
-
-    def exposed_set_channels(self, ch_list=[1]):
-        return self._module.set_ch_assignment(
-            ch_list=ch_list
-        )
-
-
-class Client(ClientBase):
-
-    def start_counting(self, bin_width=1000000000, n_bins=10000):
-        return self._service.exposed_start_counting(
-            bin_width=bin_width,
-            n_bins=n_bins
-        )
-
-    def clear_counter(self):
-        return self._service.exposed_clear_counter()
-
-    def get_counts(self):
-        res_pickle = self._service.exposed_get_counts()
-        return pickle.loads(res_pickle)
-
-    def get_x_axis(self):
-        res_pickle = self._service.exposed_get_x_axis()
-        return pickle.loads(res_pickle)
-
-    def set_channels(self, ch_list=[1]):
-        return self._service.exposed_set_channels(
-            ch_list=ch_list
-        )
-
-
-def launch(**kwargs):
-    """ Connects to SI TT and instantiates server
-
-    :param kwargs: (dict) containing relevant kwargs
-        :logger: instance of LogClient for logging purposes
-        :port: (int) port number for the Cnt Monitor server
-    """
-    DEFAULT_CH_LIST = [1]
-
-    TT.setTimeTaggerChannelNumberScheme(TT.TT_CHANNEL_NUMBER_SCHEME_ONE)
-
-    # Connect to the device, otherwise instantiate virtual connection
-    try:
-        tagger = TT.createTimeTagger()
-    except RuntimeError:
-        kwargs['logger'].warn('Failed to connect to Swabian Instruments Time Tagger.'
-                              ' Instantiating virtual device instead')
-        tagger = TT.createTimeTaggerVirtual()
-
-    cnt_trace_wrap = Wrap(
-        tagger=tagger,
-        ch_list=DEFAULT_CH_LIST,
-        logger=kwargs['logger']
-    )
-
-    # Instantiate Server
-    cnt_trace_service = Service()
-    cnt_trace_service.assign_module(module=cnt_trace_wrap)
-    cnt_trace_service.assign_logger(logger=kwargs['logger'])
-    cnt_trace_server = GenericServer(
-        service=cnt_trace_service,
-        host='localhost',
-        port=kwargs['port']
-    )
-    cnt_trace_server.start()
